backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.35.73",
    "fastapi>=0.115.0",
    "httpx>=0.28.0",
    "psycopg>=3.2.3",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.12",
    "sqlalchemy>=2.0.36",
    "surya-ocr>=0.17.1",
    "uvicorn[standard]>=0.32.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "twilio>=9.4.0",
    "transformers>=4.56.1,<5.0.0",
    "docling>=2.0.0",
    "easyocr>=1.7.0",
    "onnxruntime>=1.24.1",
]
